refactor: extract store tests
